+++ inst/doc/D_statistics.R
@@ -35,7 +35,7 @@ circular_median(san_andreas.por$azi.PoR, 1 / san_andreas$unc)
 circular_IQR(san_andreas.por$azi.PoR, 1 / san_andreas$unc)
 
 ## ----summary_stats, echo=TRUE-------------------------------------------------
-circular_summary(san_andreas.por$azi.PoR, 1 / san_andreas$unc)
+circular_summary(san_andreas.por$azi.PoR, 1 / san_andreas$unc, kappa = 10)
 
 ## ----rose1, echo=TRUE---------------------------------------------------------
 rose(san_andreas$azi,
@@ -44,17 +44,20 @@ rose(san_andreas$azi,
 )
 
 # add the density curve
-plot_density(san_andreas$azi, kappa = 10, col = "dodgerblue", shrink = 1.5)
+plot_density(san_andreas$azi, kappa = 10, col = "#51127CFF", shrink = 1.5)
 
 ## ----rose2, echo=TRUE---------------------------------------------------------
 rose(san_andreas.por$azi,
   weights = 1 / san_andreas$unc, main = "PoR",
   dots = TRUE, stack = TRUE, dot_cex = 0.5, dot_pch = 21
 )
-plot_density(san_andreas.por$azi, kappa = 10, col = "dodgerblue", shrink = 1.5)
+plot_density(san_andreas.por$azi, kappa = 10, col = "#51127CFF", shrink = 1.5)
 
 # show the predicted direction
-rose_line(135, radius = 1.1, col = "#009E73")
+rose_line(135, radius = 1.1, col = "#FB8861FF")
+
+## ----qqplot, echo=TRUE--------------------------------------------------------
+circular_qqplot(san_andreas.por$azi.PoR)
 
 ## ----random, echo=TRUE--------------------------------------------------------
 rayleigh_test(san_andreas.por$azi.PoR)