+++ NEWS.md
@@ -1,3 +1,20 @@
+# tectonicr 0.4.4
+
+# tectonicr 0.4.3.92
+* PoR_distance(): function to calculate distance to PoR
+
+# tectonicr 0.4.3
+
+* performance upgrade on spatial interpolation
+* spatial interpolation of more summary statistics
+* new vignette on spatial analysis
+* optional grid lines added to rose
+
+# tectonicr 0.4.0.9001  _2024-09-12_
+
+* bug fix in load_wsm()
+* CI now double when axial data
+
 # tectonicr 0.4.0  _2024-08-08_
 
 * minor performance upgrade